moved common tei training data code to sep module
